--- .env
@@ -1,5 +1,0 @@
-/* Database credentials */
-DB_HOST=us-cdbr-east-04.cleardb.com
-DB_USER=b75dee2cacf594
-DB_PASS=7ff4a688
-DB_NAME=heroku_5e67acb080746f5
