+++ .gitignore
@@ -1,4 +1,2 @@
 # dependencies
 /node_modules
-
-.env
