--- .gitignore.txt
@@ -1,3 +1,0 @@
-# Dependency directories
-.DS_Store
-node_modules/
